Support For Underground Stations
Adds a new undergroundStation class which represents the location of the
arrival board. This station is determined by a hard-coded station name.
With this name the code will determine the proper lines and then query
for arriving trains.

* Removes the resources directory which previously contained files used
in the font creation. (This has been moved to its own repo)
* Removed mock data when no trains are station and instead will display
a welcome message.