Support for Tfl Transport Hubs
When requesting a station that is also part of a transport hub (Old
Street, Whitechapel, etc.) the search results will provide the id for
the hub in the main API response. As a result, there was inconsistent
behavior when attempting to determine arriving trains for these
stations.

The code will now detect the stopType and then handle for stations vs.
interchanges.

* Updated README.md to include use of Dot Matrix font.
* Added additional constant for out of service message.